move dynamic func from interface for existing callers
